second

+++ src/component/Carlist.js
@@ -13,21 +13,24 @@ function Carlist() {
   }, []);
 
   const fetchCars = () => {
-    fetch('http://carrestapi.herokuapp.com/cars')
+      fetch('https://carstockrest.herokuapp.com/cars')
     .then(response => response.json())
     .then(data => setCars(data._embedded.cars))
     .catch(err => console.error(err))
   }
 
   const deleteCar = (url) => {
-    fetch(url, { method: 'DELETE' })
-    .then(response => {
-      if (response.ok)
-        fetchCars();
-      else
-        alert('Something went wrong!');
-    })
-    .catch(err => console.error(err))
+      if(window.confirm('Are you sure?')){
+          fetch(url, { method: 'DELETE' })
+              .then(response => {
+                  if (response.ok)
+                      fetchCars();
+                  else
+                      alert('Something went wrong!');
+              })
+              .catch(err => console.error(err))
+      }
+   
   }
 
   const columns = [
